+++ tests/sample_test.c
@@ -54,7 +54,7 @@ static int do_client(int argc, char *argv[])
 		printf("send error %d %d\n", ret, errno);
 		return -1;
 	}
-	printf("send '%s' on stream %ld\n", msg, sid);
+	printf("send '%s' on stream %d\n", msg, (int)sid);
 
 	flags = 0;
 	memset(msg, 0, sizeof(msg));
@@ -63,7 +63,7 @@ static int do_client(int argc, char *argv[])
 		printf("recv error %d %d\n", ret, errno);
 		return 1;
 	}
-	printf("recv '%s' on stream %ld\n", msg, sid);
+	printf("recv '%s' on stream %d\n", msg, (int)sid);
 
 	close(sockfd);
 	return 0;
@@ -122,7 +122,7 @@ static int do_server(int argc, char *argv[])
 		printf("recv error %d %d\n", ret, errno);
 		return 1;
 	}
-	printf("recv '%s' on stream %ld\n", msg, sid);
+	printf("recv '%s' on stream %d\n", msg, (int)sid);
 
 	strcpy(msg, "hello quic client!");
 	sid = QUIC_STREAM_TYPE_SERVER_MASK;
@@ -132,7 +132,7 @@ static int do_server(int argc, char *argv[])
 		printf("send error %d %d\n", ret, errno);
 		return -1;
 	}
-	printf("send '%s' on stream %ld\n", msg, sid);
+	printf("send '%s' on stream %d\n", msg, (int)sid);
 
 	close(sockfd);
 	close(listenfd);
